# alora-auth-api/docs/swagger.yaml
basePath: /
definitions:
...
  AcceptInvitationFederatedRequest:
    properties:
      token:
        example: 6f1c8a0e9d2b4f7a3c5e8b1d0a9f2c4e
        maxLength: 256
        minLength: 10
        type: string
    required:
    - token
    type: object
  AcceptInvitationRequest:
    properties:
      password:
        example: correct-horse-battery-staple
        maxLength: 512
        minLength: 8
        type: string
      token:
        example: 6f1c8a0e9d2b4f7a3c5e8b1d0a9f2c4e
        maxLength: 256
        type: string
    required:
    - password
    - token
    type: object
  AddMemberRequest:
    properties:
      email:
        example: member@acme.com
        maxLength: 320
        type: string
      user_id:
        example: 4b5c6d7e-8f90-1a2b-3c4d-5e6f7a8b9c0d
        type: string
    type: object
  AddMemberResponse:
    properties:
      group_id:
        example: 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      user_id:
        example: 4b5c6d7e-8f90-1a2b-3c4d-5e6f7a8b9c0d
        type: string
    type: object
...
  CreateInvitationRequest:
    properties:
      email:
        example: newhire@acme.com
        maxLength: 320
        type: string
      group_ids:
        example:
        - 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
        items:
          type: string
        maxItems: 20
        type: array
    required:
    - email
    type: object
  CreateInvitationResponse:
    properties:
      email:
        example: newhire@acme.com
        type: string
      expires_at:
        type: string
      id:
        example: 8c1d2e3f-4a5b-6c7d-8e9f-0a1b2c3d4e5f
        type: string
      invite_url:
        example: http://localhost:5173/accept-invitation?token=6f1c8a0e
        type: string
    type: object
...
  ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      reqId:
        example: 0b7c0f1e-6a19-4f3f-9a2d-1f9d0f3a6c21
        type: string
    type: object
...
  InvitationListItem:
    properties:
      created_at:
        type: string
      email:
        example: newhire@acme.com
        type: string
      expires_at:
        type: string
      id:
        example: 8c1d2e3f-4a5b-6c7d-8e9f-0a1b2c3d4e5f
        type: string
      status:
        enum:
        - PENDING
        - ACCEPTED
        - EXPIRED
        - REVOKED
        example: PENDING
        type: string
    type: object
  InvitationMethods:
    properties:
      google:
        example: false
        type: boolean
      password:
        example: true
        type: boolean
      sso:
        example: false
        type: boolean
    type: object
  InvitationPreview:
    properties:
      client_name:
        example: Acme Corp
        type: string
      email:
        example: newhire@acme.com
        type: string
      expires_at:
        type: string
      groups:
        example:
        - Support
        items:
          type: string
        type: array
      methods:
        $ref: '#/definitions/InvitationMethods'
    type: object
...
info:
  contact:
    name: Alora Platform
    url: https://alora.io/support
  description: |-
    App Central: one sign-in for every product of the platform, and the OpenID Provider those products trust.

    ## Two kinds of token
    People sign in once at App Central — with a password, Google, or their company's SSO, as their company's login policy allows. That opens an App Central session (an HttpOnly cookie) and yields an App Central access token (audience `app-central`), the ONLY token the `/api` routes accept. It describes everything about App Central: `scope` (what the person may do here) and `products` (the keys of the apps they may open). A product never sees it: when a product is launched, its backend runs an ordinary OpenID Connect authorization-code flow with PKCE against `/oauth/authorize` and `/oauth/token`, and receives a token for itself alone (audience `product:<key>`, carrying only its own roles).

    ## Scopes
...

    ## Group managers
...

    ## Authorizing requests here
    Sign in with `POST /auth/login/password`, copy `access_token` from the response, then press **Authorize** and paste it. Every `/api` route reads the bearer token from the `Authorization` header only — never from a cookie — and re-checks its session against the database on every request.

    ## Companies
...

    ## Errors
...
  license:
    name: Proprietary
  termsOfService: https://alora.io/terms
  title: App Central API
  version: "2.0"
paths:
...
  /api/admin/groups/{id}/members:
    post:
      consumes:
      - application/json
      description: Identify the user by EITHER `user_id` or `email`. Both the group
        and the user must belong to the company. Adding someone to the Admins group
        makes them an Admin at once.
      parameters:
      - description: Group id
        example: 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
        in: path
        name: id
        required: true
        type: string
      - description: User id or email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/AddMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/AddMemberResponse'
        "400":
          description: Neither user_id nor email supplied, or a malformed body
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Caller lacks groups:edit, lacks one of the group's scopes,
            or the person has more access
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: No such group or user in this company
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: User is already a member of this group
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a member to a group
      tags:
      - groups
  /api/admin/groups/{id}/members/{userId}:
    delete:
      description: Scoped by user, group and company together. The company's last
        active Admin cannot be removed from the Admins group.
      parameters:
      - description: Group id
        example: 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
        in: path
        name: id
        required: true
        type: string
      - description: User id
        example: 4b5c6d7e-8f90-1a2b-3c4d-5e6f7a8b9c0d
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Removed
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Caller lacks groups:edit, lacks one of the group's scopes,
            or the person has more access
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: No such membership in this company
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: The last active Admin
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a member from a group
      tags:
      - groups
...
  /api/admin/invitations:
    get:
      description: Every invitation of the company, in any status. Token hashes are
        never projected, and the invite link is not recoverable here — an invitation
        whose link was lost must be revoked and re-issued.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/InvitationListItem'
            type: array
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Caller lacks invitations:read
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List invitations
      tags:
      - invitations
    post:
      consumes:
      - application/json
      description: |-
        Issues an invitation into the company, naming the groups the new account joins — which is what gives it access (and, for the Admins group, makes it an Admin). There is no company field: it comes from the route.

        Refused if the address already belongs to a live member, or already has a live invitation. `invite_url` is returned so the link can be shared by hand when SMTP is not configured; email delivery is best-effort and off the response path.
      parameters:
      - description: Address to invite, and the groups to join
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateInvitationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateInvitationResponse'
        "400":
          description: Malformed body, or a group of another company
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Caller lacks invitations:edit, or lacks a scope one of the
            groups gives
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Address already belongs to a user, or already has a live invitation
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite a user
      tags:
      - invitations
  /api/admin/invitations/{id}:
    delete:
      description: Cancels a PENDING invitation so its link stops working. An id from
        another organisation affects no rows and reports 404.
      parameters:
      - description: Invitation id
        example: 8c1d2e3f-4a5b-6c7d-8e9f-0a1b2c3d4e5f
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Revoked
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Caller lacks invitations:edit
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: No pending invitation with that id in this company
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke an invitation
      tags:
      - invitations
...
  /auth/accept-invitation:
    post:
      consumes:
      - application/json
      description: Creates the account, adds it to the invitation's groups and consumes
        the invitation. Refused when the invitee's login policy does not allow passwords.
        Answers 204 with no body; sign in afterwards at App Central.
      parameters:
      - description: Invitation token and the password to set
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/AcceptInvitationRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Account created
        "400":
          description: Invitation not found, expired or used, passwords not allowed,
            or the password fails policy
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Accept an invitation with a password
      tags:
      - invitations
  /auth/accept-invitation/federated:
    post:
      consumes:
      - application/json
      description: Creates an account with no password, for an invitee who will sign
        in with Google or their company's SSO, adds it to the invitation's groups
        and consumes the invitation. The external identity is linked at the first
        sign-in. Refused when the invitee's login policy allows neither.
      parameters:
      - description: Invitation token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/AcceptInvitationFederatedRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Account created; sign in with Google or SSO
        "400":
          description: Invitation not found, expired or used, or neither Google nor
            SSO allowed
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: An account with this address already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Accept an invitation without a password
      tags:
      - invitations
  /auth/accept-invitation/lookup:
    get:
      description: 'The invite landing page for a visitor holding the raw token: the
        address, the company, the groups it joins, and how the invitee will sign in
        (per the login policy they will have). Unauthenticated. Unknown, expired and
        already-used tokens all return the same 400; limited to 20 attempts per 15
        minutes.'
      parameters:
      - description: Raw invitation token from the invite link
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/InvitationPreview'
        "400":
          description: Invitation not found, expired, or already used
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Preview an invitation
      tags:
      - invitations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'An App Central access token from /auth/login/password or /auth/central/refresh.
      Enter it as: Bearer &lt;token&gt;'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...